+++ lib_nn/src/c/vpu_sim.c
@@ -288,7 +288,7 @@ void VLASHR(xs3_vpu *vpu, const void *addr, const int32_t shr) {
       else if (shr >= 0)
         val = val >> shr;
       else
-        val = (unsigned)val << (-shr);
+        val = (int32_t)((uint64_t)(uint32_t)val << (-shr));
       vpu->vR.s16[i] = vpu_saturate(val, 16);
     }
   } else if (vpu->mode == MODE_S32) {
